'TestesUnitarios'

+++ src/main/java/com/springboot/crud/user/UserControler.java
@@ -1,13 +1,16 @@
 package com.springboot.crud.user;
 
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 import javax.validation.Valid;
+
 
 
 @Controller
@@ -19,54 +22,54 @@ public class UserControler {
         this.userRepository = userRepository;
     }
 
-    @RequestMapping(value ="/signup",method = RequestMethod.GET)
+    /*@RequestMapping(value ="/signup",method = RequestMethod.GET)
     public String showSignUpForm(User user) {
-        return "index";
+        return "add-user";
     }
-
+    */
     @RequestMapping(value ="/adduser", method = RequestMethod.POST)
-    public String addUser(@Valid User user, BindingResult result, Model model) {
-        if (result.hasErrors()) {
-            return "add-user";
-        }
-
+    public ResponseEntity addUser(@RequestBody @Valid User user, BindingResult result, Model model) {
         userRepository.save(user);
-        return "redirect:/index";
+        return ResponseEntity.ok(user);
     }
 
-    @RequestMapping(value ="/index",method = RequestMethod.GET)
-    public String showUserList(Model model) {
-        model.addAttribute("users", userRepository.findAll());
-        return "index";
-    }
-
-
-    @RequestMapping(value ="/edit/{id}", method = RequestMethod.GET)
-    public String showUpdateForm(@PathVariable("id") long id, Model model) {
+    @RequestMapping(value ="/read/{id}", method = RequestMethod.GET)
+    public ResponseEntity<User> showUpdateForm(@PathVariable("id") int id, Model model) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
-
-        model.addAttribute("user", user);
-        return "update-user";
+                .orElse(null);;
+        if (user != null) {
+            model.addAttribute("user", user);
+            return ResponseEntity.ok(user);
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @RequestMapping(value="/update/{id}",method = RequestMethod.PATCH)
-    public String updateUser(@PathVariable("id") long id, @Valid User user,
-                             BindingResult result, Model model) {
-        if (result.hasErrors()) {
-            user.setId(id);
-            return "update-user";
+    public ResponseEntity updateUser(@PathVariable("id") int id, @RequestBody @Valid User user, BindingResult result, Model model) {
+        User UserEdit = userRepository.findById(id)
+                .orElse(null);;
+        if (UserEdit !=null) {
+            UserEdit.setName(user.getName());
+            UserEdit.setEmail(user.getEmail());
+            return ResponseEntity.ok().body(userRepository.save(UserEdit));
         }
-
-        userRepository.save(user);
-        return "redirect:/index";
+        return ResponseEntity.notFound().build();
     }
 
-    @RequestMapping(value="/delete/{id}",method = RequestMethod.DELETE)
-    public String deleteUser(@PathVariable("id") long id, Model model) {
+    @RequestMapping(value="/delete/{id}",method = RequestMethod.GET)
+    public ResponseEntity<User> deleteUser(@PathVariable("id") int id, Model model) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
-        userRepository.delete(user);
-        return "redirect:/index";
+                .orElse(null);;
+        if (user != null){
+            userRepository.delete(user);
+            return ResponseEntity.ok(user);
+        }
+        return ResponseEntity.notFound().build();
+    }
+
+    @RequestMapping(value ="/getall",method = RequestMethod.GET)
+    public ResponseEntity<Iterable<User>> showUserList(Model model) {
+        model.addAttribute("users", userRepository.findAll());
+        return ResponseEntity.ok(userRepository.findAll());
     }
 }
